# server/openapi.yaml
openapi: 3.0.3
info:
  title: WEBonTour API
  description: API for the WEBonTour CMS
  version: 1.0.0
  contact:
    name: Project Team
  license:
    name: Project License
servers:
  - url: /api
    description: API base path
...
paths:
...
  /posts:
    get:
      tags:
        - posts
      summary: Get all posts
      description: Get a list of all posts with optional filtering
      operationId: getPosts
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category name
        - name: tag
          in: query
          schema:
            type: string
          description: Filter by tag name
        - name: author
          in: query
          schema:
            type: string
          description: Filter by author username
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
          description: Number of posts to return
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
          description: Number of posts to skip
      responses:
        '200':
          description: Posts retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - posts
      summary: Create a new post
      description: Create a new blog post (writer or admin)
      operationId: createPost
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - title
                - content
                - category_id
              properties:
                title:
                  type: string
                  example: My New Blog Post
                content:
                  type: string
                  example: This is the content of my blog post
                category_id:
                  type: integer
                  example: 1
                tags:
                  type: array
                  items:
                    type: integer
                  example: [1, 2, 3]
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Attachments for the post
      responses:
        '201':
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Bad request - validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized (writer or admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /document/{filename}:
    get:
      tags:
        - documents
      summary: Get document
      description: Get a legal document or policy by filename
      operationId: getDocument
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
          description: Document filename without extension
          example: privacy
      responses:
        '200':
          description: Document content
          content:
            text/markdown:
              schema:
                type: string
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Post:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: My Blog Post
        content:
          type: string
          example: This is the content of my blog post...
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        author:
          type: string
          example: johndoe
        author_id:
          type: integer
          example: 1
        category_id:
          type: integer
          example: 1
        category_name:
          type: string
          example: Technology
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
...
    Tag:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: JavaScript
    Attachment:
      type: object
      properties:
        id:
          type: integer
          example: 1
        filename:
          type: string
          example: file-1620000000000-123456789.jpg
    Error:
      type: object
      properties:
        message:
          type: string
          example: Error message